break up the components into separate folders
