+++ app/index.js
@@ -1,7 +1,7 @@
 const React = require('react');
 const ReactDOM = require('react-dom');
 require('./style.css');
-const App = require('./components/App');
+const App = require('./components/App/App');
 
 ReactDOM.render(
   <App />,
